+++ lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:gemini_app/presentation/gemini_app.dart';
+
+import './presentation/select_photo_page.dart';
 
 void main() {
-  runApp(const GeminiApp());
+  runApp(const SelectPhotoPage());
 }
